Update validation for telemetry data

+++ f1data/services/telemetry/models/Telemetry.py
@@ -6,11 +6,11 @@ from pydantic import BaseModel, ConfigDict, field_serializer
 class TelemetryData(BaseModel):
     model_config = ConfigDict(arbitrary_types_allowed=True)
 
-    Throttle: Sequence[int]
+    Throttle: Sequence[float]
     Brake: Sequence[Boolean]
     Gear: Sequence[int]
-    Speed: Sequence[int]
-    RPM: Sequence[int]
+    Speed: Sequence[float]
+    RPM: Sequence[float]
     Time: Sequence[Timedelta]
     RelativeDistance: Sequence[float]
     Distance: Sequence[float]
